refactor: QnAHome의 상태관리 코드 custom hook(useQnAHomeData.ts)으로 분리
